+++ internal/delivery/http/request/exchange_request/update_keys.go
@@ -7,11 +7,11 @@ import (
 type KeyData struct {
 	Name  string `json:"name" validate:"required,min=1,max=255"`
 	Value string `json:"value" validate:"required,min=1,max=255"`
-} // @name ExchangeUpdateKeyData
+} //	@name	ExchangeUpdateKeyData
 
 type UpdateKeys struct {
 	Keys []KeyData `json:"keys" required:"true" validate:"required,dive"`
-} // @name ExchangeUpdateKeysRequest
+} //	@name	ExchangeUpdateKeysRequest
 
 func (uk *UpdateKeys) ToMap() map[models.ExchangeKeyName]*string {
 	res := make(map[models.ExchangeKeyName]*string, len(uk.Keys))