+++ internal/delivery/http/request/store_api_key_request/update.go
@@ -2,4 +2,4 @@ package store_api_key_request
 
 type UpdateStatusRequest struct {
 	Status bool `db:"status" json:"status" validate:"boolean"`
-} // @name UpdateAPIKeyStatusRequest
+} //	@name	UpdateAPIKeyStatusRequest