+++ internal/delivery/http/request/store_webhook_request/update.go
@@ -8,4 +8,4 @@ type UpdateRequest struct {
 	URL     string                 `db:"url" json:"url" validate:"required,http_url"`
 	Enabled bool                   `db:"enabled" json:"enabled"  validate:"boolean"`
 	Events  []*models.WebhookEvent `db:"events" json:"events" validate:"required,dive"`
-} // @name UpdateStoreWebhookRequest
+} //	@name	UpdateStoreWebhookRequest