+++ internal/delivery/http/request/store_whitelist_request/update.go
@@ -2,4 +2,4 @@ package store_whitelist_request
 
 type UpdateRequest struct {
 	Ips []string `db:"ips" json:"ips" validate:"required,dive,ip" format:"ipv4"`
-} // @name UpdateWhitelistRequest
+} //	@name	UpdateWhitelistRequest