+++ internal/delivery/http/responses/search_response/search_by_criteria.go
@@ -3,7 +3,7 @@ package search_response
 type SearchByCriteriaResponse[T any] struct {
 	SearchType string `json:"search_type"`
 	Data       T      `json:"data"`
-} // @name SearchByCriteriaResponse
+} //	@name	SearchByCriteriaResponse
 
 func PrepareByCriteria[T any](searchType string, data T) SearchByCriteriaResponse[T] {
 	return SearchByCriteriaResponse[T]{