+++ internal/delivery/http/responses/wallet_response/public.go
@@ -1,3 +1,3 @@
 package wallet_response
 
-type GetPublicWalletDataResponse struct{} // @name PublicWalletDataResponse
+type GetPublicWalletDataResponse struct{} //	@name	PublicWalletDataResponse