+++ internal/storage/storecmn/params.go
@@ -4,7 +4,7 @@ type CommonFindParams struct {
 	IsAscOrdering bool   `json:"is_asc_ordering"`
 	OrderBy       string `json:"order_by"`
 	PageParams
-} // @name CommonFindParams
+} //	@name	CommonFindParams
 
 func NewCommonFindParams() *CommonFindParams {
 	return &CommonFindParams{}
@@ -33,21 +33,21 @@ func (s *CommonFindParams) SetPageSize(v *uint32) *CommonFindParams {
 type PageParams struct {
 	Page     *uint32
 	PageSize *uint32
-} // @name PageParams
+} //	@name	PageParams
 
 type FindResponseWithPagingFlag[T any] struct {
 	Items            []T  `json:"items"`
 	IsNextPageExists bool `json:"is_next_page_exists"`
-} // @name ResponseWithPagingFlag
+} //	@name	ResponseWithPagingFlag
 
 type FullPagingData struct {
 	Total    uint64 `json:"total"`
 	PageSize uint64 `json:"page_size"`
 	Page     uint64 `json:"page"`
 	LastPage uint64 `json:"last_page"`
-} // @name FullPagingData
+} //	@name	FullPagingData
 
 type FindResponseWithFullPagination[T any] struct {
 	Items      []T            `json:"items"`
 	Pagination FullPagingData `json:"pagination"`
-} // @name ResponseWithFullPagination
+} //	@name	ResponseWithFullPagination