+++ CHANGELOG.md
@@ -6,8 +6,12 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
-
 - Currency reference data about it being native, and contract address for external routes [DV-3761]
+- fix: duplicates withdrawal from processing on 'in_progress' transfer stage [DV-3933]
+
+## [0.9.14] - 2025-13-04
+- feat: added extended tx info in search response [DV-3827]
+- feat: added endpoint to retrieve extended information about store enabled currencies [DV-3857]
 
 ## [0.9.13] - 2025-11-19
 - Fix `FindLastWalletTransactions` sql query  [DV-3808]